+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [0.2.4]
+
+### Added
+- Add a simple after-combat dialog to assign monster parts to defeated NPCs
+- Add weakness to the list of extended actor roll options
+
+### Changed
+- Change all actor roll options to start with self
+
+### Fixed
+- Fix Wisdom imbuement not applying to Perception items
+
 ## [0.2.3] - 2025-09-29
 
 ### Fixed